tests/krate/publish: Extract `tarball` module
